One parcel config to rule them all
