+++ packages/core/core/src/PackagerRunner.js
@@ -37,6 +37,7 @@ import {PARCEL_VERSION, HASH_REF_PREFIX, HASH_REF_REGEX} from './constants';
 
 type Opts = {|
   config: ParcelConfig,
+  configRef?: number,
   farm?: WorkerFarm,
   options: ParcelOptions,
   optionsRef?: number,
@@ -59,6 +60,7 @@ const BOUNDARY_LENGTH = HASH_REF_PREFIX.length + 32 - 1;
 
 export default class PackagerRunner {
   config: ParcelConfig;
+  configRef: ?number;
   options: ParcelOptions;
   optionsRef: ?number;
   farm: ?WorkerFarm;
@@ -69,13 +71,14 @@ export default class PackagerRunner {
   getBundleInfoFromWorker: ({|
     bundle: InternalBundle,
     bundleGraphReference: number,
-    config: ParcelConfig,
+    configRef: number,
     cacheKeys: CacheKeyMap,
     optionsRef: number,
   |}) => Promise<BundleInfo>;
 
-  constructor({config, farm, options, optionsRef, report}: Opts) {
+  constructor({config, configRef, farm, options, optionsRef, report}: Opts) {
     this.config = config;
+    this.configRef = configRef;
     this.options = options;
     this.optionsRef = optionsRef;
     this.pluginOptions = new PluginOptions(this.options);
@@ -151,7 +154,7 @@ export default class PackagerRunner {
         bundleGraphReference,
         cacheKeys,
         optionsRef: nullthrows(this.optionsRef),
-        config: this.config,
+        configRef: nullthrows(this.configRef),
       }));
 
     return {time: Date.now() - start, hash, hashReferences, cacheKeys};
