+++ packages/core/core/src/RequestTracker.js
@@ -296,6 +296,13 @@ export default class RequestTracker {
     this.graph.removeById(id);
   }
 
+  storeResult(id: string, result: mixed) {
+    let node = this.graph.getNode(id);
+    if (node && node.type === 'request') {
+      node.value.result = result;
+    }
+  }
+
   hasValidResult(id: string) {
     return (
       this.graph.nodes.has(id) &&
@@ -362,6 +369,7 @@ export type RequestRunnerAPI = {|
   invalidateOnFileUpdate: FilePath => void,
   invalidateOnStartup: () => void,
   replaceSubrequests: (Array<RequestGraphNode>) => void,
+  storeResult: (result: mixed) => void,
 |};
 
 export function generateRequestId(type: string, request: mixed) {
@@ -415,9 +423,7 @@ export class RequestRunner<TRequest, TResult> {
   // unused vars are used for types
   // eslint-disable-next-line no-unused-vars
   onComplete(request: TRequest, result: TResult, api: RequestRunnerAPI) {
-    throw new Error(
-      `RequestRunner for type ${this.type} did not implement onComplete()`,
-    );
+    // Do nothing, this is defined for flow if extended classes implement this function
   }
 
   generateRequestId(request: TRequest) {
@@ -436,6 +442,9 @@ export class RequestRunner<TRequest, TResult> {
         this.tracker.graph.invalidateOnStartup(requestId),
       replaceSubrequests: subrequestNodes =>
         this.tracker.graph.replaceSubrequests(requestId, subrequestNodes),
+      storeResult: result => {
+        this.tracker.storeResult(requestId, result);
+      },
     };
 
     return api;
