+++ packages/core/core/test/AssetGraphBuilder.test.js
@@ -1,22 +1,15 @@
 // @flow strict-local
 
 import invariant from 'assert';
-import path from 'path';
-import nullthrows from 'nullthrows';
 
 import {createWorkerFarm} from '../';
 import AssetGraphBuilder from '../src/AssetGraphBuilder';
-import {resolveParcelConfig} from '../src/loadParcelConfig';
 import {DEFAULT_OPTIONS} from './utils';
-
-const FIXTURES_DIR = path.join(__dirname, 'fixtures');
-const CONFIG_DIR = path.join(FIXTURES_DIR, 'config');
 
 describe('AssetGraphBuilder', function() {
   // This depends on spinning up a WorkerFarm, which can take some time.
   this.timeout(20000);
 
-  let config;
   let builder;
   let workerFarm;
 
@@ -29,19 +22,11 @@ describe('AssetGraphBuilder', function() {
   });
 
   beforeEach(async () => {
-    config = nullthrows(
-      await resolveParcelConfig(
-        path.join(CONFIG_DIR, 'index'),
-        DEFAULT_OPTIONS,
-      ),
-    ).config;
-
     builder = new AssetGraphBuilder();
     await builder.init({
       name: 'test',
       options: DEFAULT_OPTIONS,
       optionsRef: 1,
-      config,
       entries: ['./module-b'],
       workerFarm,
     });
