+++ packages/core/integration-tests/test/integration/custom-config/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "parcel-integration-custom-config",
+  "version": "1.2.3"
+}
